Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,7 @@
-#include <stdio.h>
+#include<stdio.h>
 int main(){
     int a, b;
-    float res;
+    int res;
     char c;
     scanf("%d %d %c", &a, &b, &c);
     switch(c){
